+++ packages/react/package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "deploy": "pnpm build && pnpm upload",
-    "upload": "pnpm upload-bundle ./dist testnet partfs1.testnet gormp.testnet",
+    "upload": "pnpm upload-bundle ./dist testnet partfs1.testnet gormp.testnet react-vite-example",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
@@ -33,5 +33,8 @@
     "rollup-plugin-visualizer": "^5.12.0",
     "typescript": "^5.2.2",
     "vite": "^5.2.0"
+  },
+  "bosConfig": {
+    "application": "react-vite-example"
   }
 }
