+++ packages/react/vite.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig, type PluginOption } from 'vite'
+import { replaceHtmlPaths } from '@chain-deployed-ui/presets';
 import { presetBundles } from '@chain-deployed-ui/presets/react';
 import react from '@vitejs/plugin-react'
 import gzipPlugin from 'rollup-plugin-gzip';
@@ -10,8 +11,7 @@ export default defineConfig({
     name: 'inject-bundle-url',
     apply: 'build',
     transformIndexHtml(html ) {
-      // TODO prefix root-level assets - this only removes the leading /
-      return html.replace(/(href|src)="\/((?:[\w\s-]+\/)*[^/]+\.(?:js|css|svg))/g, `$1="$2`);
+      return replaceHtmlPaths(html);
     }
   }],
   define: {
